+++ lib/emerge_cli.rb
@@ -34,7 +34,9 @@ require_relative 'utils/profiler'
 require_relative 'utils/project_detector'
 require_relative 'utils/macho_parser'
 require_relative 'utils/version_check'
-require_relative 'utils/xcode_device'
+require_relative 'utils/xcode_device_manager'
+require_relative 'utils/xcode_simulator'
+require_relative 'utils/xcode_physical_device'
 
 require 'dry/cli'
